Rename endpointcertificates -> certificate
